изменил структуру приложения
