+++ src/App.tsx
@@ -3,8 +3,8 @@ import './App.css';
 import Header from './components/Header/Header';
 import Navbar from "./components/Navbar/Navbar";
 import Profile from './components/Profile/Profile'
-import Dialogs from "./components/Dialogs/Dialogs";
 import {BrowserRouter, Route} from "react-router-dom";
+import Dialogs from "./components/Dialogs/Dialogs";
 
 const App = () => {
     return (
