+++ src/components/Dialogs/Message/Message.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import s from '../Dialogs.module.css';
+
+
+
+type MessageType = {
+    message: string
+}
+
+export const Message = (props: MessageType) =>{
+    return  <div className={s.message}>{props.message}</div>
+}
+
+
